requirements.txt
dnspython
requests
pillow
pandas